add a little test
