+++ src/Frame2.java
@@ -4,6 +4,8 @@ import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.awt.geom.AffineTransform;
 import java.awt.image.BufferedImage;
+import java.util.ArrayList;
+import java.util.List;
 
 public class Frame2 extends DefaultFrame {
     private JButton selectedImagePanelButton;
@@ -84,13 +86,16 @@ public class Frame2 extends DefaultFrame {
 
         @Override
         public void actionPerformed(ActionEvent e) {
-                selectedImagePanelButton = mainPanelButton;
+            selectedImagePanelButton = mainPanelButton;
         }
     }
 
-    private class ImagePanelButtonListener implements ActionListener {
+    public class ImagePanelButtonListener implements ActionListener {
         private final JButton imagePanelButton;
         private final int buttonIndex;
+
+        // Array to store clicked button indexes (instance variable)
+        private List<Integer> clickedButtonIndexes = new ArrayList<>();
 
         public ImagePanelButtonListener(JButton imagePanelButton, int buttonIndex) {
             this.imagePanelButton = imagePanelButton;
@@ -119,23 +124,39 @@ public class Frame2 extends DefaultFrame {
                         selectedImagePanelButton.setIcon(scaledIcon);
                         // Set the icon of the clicked button in the image panel to null (remove the icon)
                         imagePanelButton.setIcon(null);
+
+                        // Update the array with the clicked button index
+                        updateClickedButtonIndexes(buttonIndex);
                     }
 
                 } else {
-                    if (selectedImagePanelButton.getIcon() != null){
+                    if (selectedImagePanelButton.getIcon() != null) {
                         // If it was blank, restore the initial image of the clicked button in the image panel
                         imagePanelButton.setIcon(initialImageIcons[buttonIndex]);
 
                         // Reset the icon of the clicked button in the main panel to blank
                         selectedImagePanelButton.setIcon(null);
-                    }
 
+                        // Update the array with the clicked button index
+                        updateClickedButtonIndexes(buttonIndex);
+                    }
                 }
 
                 selectedImagePanelButton = null;
             }
         }
+
+        // Method to update the array with the clicked button index
+        private void updateClickedButtonIndexes(int buttonIndex) {
+            if (!clickedButtonIndexes.contains(buttonIndex)) {
+                clickedButtonIndexes.add(buttonIndex);
+            }
+
+            // Print the array for demonstration purposes
+            System.out.println("Clicked Button Indexes: " + clickedButtonIndexes);
+        }
     }
+
 
     private String getImagePathFromButton(JButton button) {
         // Assuming that the image path is stored as the action command of the button
@@ -176,51 +197,5 @@ public class Frame2 extends DefaultFrame {
         return button;
     }
 
-    private class RotatePanelButtonListener implements ActionListener {
-        private final int buttonNumber;
-        private final JPanel mainPanel;
 
-        public RotatePanelButtonListener(int buttonNumber, JPanel mainPanel) {
-            this.buttonNumber = buttonNumber;
-            this.mainPanel = mainPanel;
-        }
-
-        @Override
-        public void actionPerformed(ActionEvent e) {
-            // Rotate the image of the corresponding button in the main panel
-            for (Component component : mainPanel.getComponents()) {
-                if (component instanceof JButton) {
-                    JButton mainPanelButton = (JButton) component;
-                    if (Integer.parseInt(mainPanelButton.getName()) == buttonNumber) {
-                        rotateButtonImage(mainPanelButton);
-                        break;
-                    }
-                }
-            }
-        }
-
-        private void rotateButtonImage(JButton button) {
-            // Get the icon from the button
-            ImageIcon buttonIcon = (ImageIcon) button.getIcon();
-
-            if (buttonIcon != null) {
-                // Rotate the image
-                Image rotatedImage = rotateImage(buttonIcon.getImage(), Math.toRadians(90));
-                button.setIcon(new ImageIcon(rotatedImage));
-            }
-        }
-
-        private Image rotateImage(Image image, double angle) {
-            AffineTransform transform = new AffineTransform();
-            transform.rotate(angle, image.getWidth(null) / 2.0, image.getHeight(null) / 2.0);
-
-            BufferedImage rotatedImage = new BufferedImage(image.getWidth(null), image.getHeight(null), BufferedImage.TRANSLUCENT);
-            Graphics2D g = rotatedImage.createGraphics();
-            g.setTransform(transform);
-            g.drawImage(image, 0, 0, null);
-            g.dispose();
-
-            return rotatedImage;
-        }
-    }
 }
