+++ src/Main.java
@@ -1,67 +1,12 @@
+import javax.swing.*;
+import java.util.ArrayList;
+import java.util.List;
+
 public class Main {
+    private static JButton submitButton;
+    private static List<Integer> storedButtonIndexes = new ArrayList<>();
     public static void main(String[] args) {
         Frame1 frame1 = new Frame1();
-//        Frame2 frame2 = new Frame2();
-        System.out.println("Hello world!");
-        Card_1 card_1 = new Card_1();
-        Card_2 card_2 = new Card_2();
-        Card_3 card_3 = new Card_3();
-        Card_4 card_4 = new Card_4();
-        Card_5 card_5 = new Card_5();
-        Card_6 card_6 = new Card_6();
-        Card_7 card_7 = new Card_7();
-        Card_8 card_8 = new Card_8();
-        Card_9 card_9 = new Card_9();
-
-
-//        // Test case 1, 2 :
-//        card_1.rotateRight();
-//        card_1.rotateRight();
-//        card_1.rotateRight();
-//
-//        card_2.rotateRight();
-//
-//        card_3.rotateRight();
-//        card_3.rotateRight();
-//        card_3.rotateRight();
-//
-//        card_4.rotateRight();
-//        card_4.rotateRight();
-//        card_4.rotateRight();
-//
-//        card_6.rotateRight();
-//
-//        card_7.rotateRight();
-//        card_7.rotateRight();
-//        card_7.rotateRight();
-//
-//        card_8.rotateRight();
-//
-//        // used for test case 1: true
-//        Card[][] board = {
-//            {card_1, card_6, card_5},
-//            {card_2, card_4, card_8},
-//            {card_7, card_9, card_3}
-//        };
-//
-//        // used for test case 2: true
-////        Card[][] board = {
-////            {card_1, card_6, card_5},
-////            {card_3, card_4, card_8},
-////            {card_7, card_9, card_2}
-////        };
-
-        // Test case 3: false
-        Card[][] board = {
-                {card_7, card_6, card_4},
-                {card_8, card_3, card_9},
-                {card_1, card_2, card_5}
-        };
-
-        Checkfunction checkfunction = new Checkfunction();
-        checkfunction.printSquaredMatrix(board);
-
-        boolean checkSolution = checkfunction.checkSolution(board);
-        System.out.println(checkSolution);
+        
     }
 }
